Moved VideoExporter to visualiser/ directory
