+++ CMakeLists.txt
@@ -14,10 +14,10 @@ file(GLOB Sources *.cpp)
 list(APPEND Sources 
     config/Config.cpp 
     visualiser/SettingParameter.cpp
+    visualiser/VideoExporter.cpp
     widgets/ClickableLabel.cpp
     widgets/ConfigDetailsDialog.cpp
     widgets/SceneWidget.cpp
-    video/VideoExporter.cpp
 )
 file(GLOB Resources *.qrc)
 file(GLOB Styles *.qss)
